+++ src/core/domain/server/usecases/users-metadata/add-user-metadata.interface.ts
@@ -1,4 +1,4 @@
-import { HttpStatusCodeEnum } from '~/core/app/server';
+import { HttpStatusCodeEnum } from '~/core/app/server/protocols';
 import { UserMetadataCategoryEnum } from './add-user-metadata.enum';
 
 export interface AddUserMetadata {
